requirements.txt
beautifulsoup4
huggingface_hub
python-telegram-bot
requests
tqdm